+++ elastic/requirements-dev.txt
@@ -1,3 +1,1 @@
-mock==2.0.0
-pytest
-pytest-benchmark
+datadog-checks-dev
