+++ elastic/tests/compose/elastic.yaml
@@ -7,6 +7,7 @@ services:
       - cluster.name=docker-cluster
       - bootstrap.memory_lock=true
       - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
+      - ${ELASTIC_CONFIG}
     ports:
       - "9200:9200"
       - "9300:9300"
@@ -14,8 +15,6 @@ services:
       memlock:
         soft: -1
         hard: -1
-    # volumes:
-    #  - esdata1:/usr/share/elasticsearch/data
     networks:
       - network1
