+++ elastic/tox.ini
@@ -83,6 +83,11 @@ setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=6.2.4
 
+[testenv:elastic6.3]
+setenv =
+    ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
+    ELASTIC_VERSION=6.3.2
+
 [testenv:bench]
 setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
